refactor(posts-tag-manager): add post-list-manager to manage post list
