--- packages/mirror-tv-next/styles/components/tag/more-posts-list.module.scss
@@ -1,20 +1,0 @@
-@import '/styles/theme/breakpoints';
-
-.list,
-.btnWrapper {
-  margin-top: 28px;
-  @include breakpoint(md) {
-    margin: 48px 0 0 0;
-  }
-  @include breakpoint(xl) {
-    margin: 40px 0 0 0;
-  }
-  @include breakpoint(xxl) {
-    margin: 53px 0 0 0;
-  }
-}
-
-.btnWrapper {
-  display: flex;
-  justify-content: center;
-}
