+++ packages/mirror-tv-next/styles/components/tag/posts-list-manager.module.scss
@@ -1,5 +1,24 @@
 @import '/styles/theme/breakpoints';
 
+.list,
+.btnWrapper {
+  margin-top: 28px;
+  @include breakpoint(md) {
+    margin: 48px 0 0 0;
+  }
+  @include breakpoint(xl) {
+    margin: 40px 0 0 0;
+  }
+  @include breakpoint(xxl) {
+    margin: 53px 0 0 0;
+  }
+}
+
+.btnWrapper {
+  display: flex;
+  justify-content: center;
+}
+
 .posts {
   list-style: none;
   display: grid;
